move searchUsingGrpcApi to utils
